This initial commit adds a subcommand to remove default VPCs from an account.
One complexity introduced by this change is that tests will break as they rely
on having default VPCs available in which to create resources, but the tests for
this functionality intentionally removes all the VPCs. Therefore, this will need
to be refactored to use the mock harnesses (ec2iface) provided by aws-sdk-go.

Nonetheless, the basic functionality is very close to working.